requirements.txt
flask
pillow
requests
visvis